Separate simplifier in its own file
